+++ adapters/pixml.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
 # (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.rst.
-# version: 0.20.6-2-g890f5fc
+
 from __future__ import print_function
 from __future__ import unicode_literals
 from __future__ import absolute_import
@@ -13,7 +13,7 @@ import copy
 import datetime
 import numpy as np
 import re
-import sys
+
 
 class Series(object):
     """
@@ -69,15 +69,15 @@ class Series(object):
             time=self.end.strftime('%H:%M:%S'),
         )
         self._find(self.header, 'parameterId').text = name
-        
+
     def _index(self, dt):
-        span = dt - self.start 
+        span = dt - self.start
         step = self.step
         return int(span.total_seconds() / step.total_seconds())
 
     def _datetime_from_index(self, index):
         return self.start + index * self.step
-        
+
     def _tag(self, tag):
         """ Return namespaced tag. """
         return self.TAG.format(self.ns, tag)
@@ -90,7 +90,7 @@ class Series(object):
         if ns is None:
             ns = self._ns(elem)  # Assume same namespace
         return elem.find(self.TAG.format(ns, tag))
-        
+
 
 class SeriesReader(object):
     PATTERN = re.compile('\{([^{}]*)\}([^{}]*)')
@@ -98,7 +98,6 @@ class SeriesReader(object):
 
     def __init__(self, xml_input_path):
         self.xml_input_path = xml_input_path
-
 
     def _datetime_from_elem(self, elem):
         """ Return python datetime object. """
@@ -116,7 +115,6 @@ class SeriesReader(object):
             '{}s'.format(elem.attrib['unit']): int(elem.attrib['multiplier']),
         }
         return datetime.timedelta(**td_kwargs)
-
 
     def _series_from_tree(self, tree):
         """
@@ -149,9 +147,11 @@ class SeriesReader(object):
         """
         Returns a generator of series objects.
 
-        As etree parses the tree, it keeps an internal tree that is not necessarily in sync with the events returned by iterparse.
+        As etree parses the tree, it keeps an internal tree that is not
+        necessarily in sync with the events returned by iterparse.
 
-        Therefore we keep a copy of selected elements of the tree that is used to instantiate the series.
+        Therefore we keep a copy of selected elements of the tree that
+        is used to instantiate the series.
         """
         iterator = iter(ElementTree.iterparse(
             self.xml_input_path, events=('start', 'end'),
@@ -218,15 +218,18 @@ class SeriesWriter(object):
         map(header.remove, elements)
 
         # Add range elements
-        time = ElementTree.SubElement(header, 'timeStep',
-            unit='second', 
+        time = ElementTree.SubElement(
+            header, 'timeStep',
+            unit='second',
             multiplier=str(series.step.seconds),
         )
-        start = ElementTree.SubElement(header, 'startDate', 
+        start = ElementTree.SubElement(
+            header, 'startDate',
             date=series.start.strftime('%Y-%m-%d'),
             time=series.start.strftime('%H:%M:%S'),
         )
-        stop = ElementTree.SubElement(header, 'endDate',
+        stop = ElementTree.SubElement(
+            header, 'endDate',
             date=series.stop.strftime('%Y-%m-%d'),
             time=series.stop.strftime('%H:%M:%S'),
         )
@@ -234,21 +237,21 @@ class SeriesWriter(object):
         # Correct indentation
         for element in time, start, stop:
             element.tail = elements[0].tail
-         
+
         # Add remaining header elements
         header.extend(elements)
 
     def _write_tree(self, tree, begin=None, end=None, indent=0):
-        """ 
+        """
         Write a part of the tree stringlist.
 
         Begin and end are strings that mark begin and end of the part of
         the tree that needs to be written. May not work if elementtrees
         splitting behaviour gets really weird.
         """
-        
+
         self.xml_output_file.write(indent * ' ')
-        
+
         write = True if begin is None else False
 
         for text in ElementTree.tostringlist(tree):
@@ -265,9 +268,9 @@ class SeriesWriter(object):
             self.xml_output_file.write('\n')
 
     def _write_series(self, series):
-        """ 
-        Write series to xmlfile. 
-        
+        """
+        Write series to xmlfile.
+
         Event elements are generated and written one by one to keep
         memory consumption low.
         """
@@ -278,7 +281,7 @@ class SeriesWriter(object):
         # Complete the header and write it
         self._add_range_elements(tree, series)
         self._write_tree(tree, begin='<series', end='</header>', indent=4)
-        
+
         # Write the events
         for dt, value in series:
             self._write_flat_element(
@@ -311,84 +314,3 @@ class SeriesWriter(object):
             self._write_tree(tree, begin='</TimeSeries')
 
         self.xml_output_file.close()
-
-
-class PercentileConverter(object):
-    
-    PERIODS = {
-        '10w': 10,
-        '6m': 26,
-        '1j': 52,
-    }
-    PERCENTILES = (10, 50, 90)
-    PARAMETERS = {}
-    for k, v in PERIODS.items():
-        for p in PERCENTILES:
-            parameterkey = 'Q.%s.%s' % (p, k)
-            PARAMETERS.update({
-                parameterkey: {'percentile': p, 'period': v},
-            })
-
-    def _percentile_series(self, series):
-        """
-        """
-        # Create and fill an array with shape (weeks, steps-per-week)
-        width = int(3600 * 24 * 7 / series.step.total_seconds())
-        height = int(np.ceil(len(series) / width))
-        size = height * width
-        table = np.append(
-            series.ma[::-1],
-            np.ma.array(
-                np.zeros(size - len(series)),
-                mask=True,
-                fill_value=series.ma.fill_value,
-            )
-        )
-        table.shape = height, width
-
-        for name, parameter in self.PARAMETERS.iteritems():
-            if parameter['period'] > height:
-                continue
-
-            ma = np.percentile(
-                table[0:parameter['period']],
-                parameter['percentile'],
-                axis=0,
-            )[::-1]
-            stop = series.stop
-            step = series.step
-            start = stop - step * (width - 1)
-            tree = copy.deepcopy(series.tree)
-            for elem in tree.iter():
-                if elem.tag.endswith('parameterId'):
-                    elem.text = name
-
-            yield Series(tree=tree, start=start, stop=stop, step=step, ma=ma)
-        
-    def convert(self, series_iterable):
-        for series in series_iterable:
-            for result in self._percentile_series(series):
-                yield(result)
-
-
-def percentiles(*args):
-    """
-    To be called by the console script.
-    
-    Since it is not sure if the console script passes the arguments,
-    we accept them, but don't use them.
-    """
-    xml_input_path = sys.argv[1]
-    xml_output_path = sys.argv[2]
-
-    reader = SeriesReader(xml_input_path)
-    converter = PercentileConverter()
-    writer = SeriesWriter(xml_output_path)
-
-    writer.write(converter.convert(reader.read()))
-    
-    return 0
-
-
-if __name__ == '__main__':
-    exit(percentiles())